requirements.txt
gradio>=4.44.0
openai>=1.12.0
python-dotenv>=1.0.0
pandas>=2.0.0
uvicorn>=0.24.0
fastapi>=0.104.0
pydantic>=2.0.0
aiohttp>=3.9.0
requests>=2.31.0
azure-cognitiveservices-speech>=1.34.0
plotly>=5.17.0
kaleido>=0.2.1

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-html>=4.1.0
pytest-json-report>=1.5.0
pytest-xdist>=3.3.0
coverage>=7.3.0
mock>=5.1.0
